test: move tests to test/server/socket-injection.test
